Get rid of executor in the consumer
Kafka has an ordering guarantee for partitions. By submitting incoming
messages into a thread pool we lose this ordering guarantee. To keep the
ordering guarantee and also keep the implementation simple consumer are
not single threaded.
