+++ example/avro.rb
@@ -17,6 +17,8 @@ end
 
 
 GlassOctopus.run(app) do |config|
+  config.logger = Logger.new(STDOUT)
+
   config.adapter :ruby_kafka do |kafka|
     kafka.broker_list = array_from_env("KAFKA_BROKER_LIST", default: %w[localhost:9092])
     kafka.topic       = ENV.fetch("KAFKA_TOPIC", "mytopic")
