+++ example/basic.rb
@@ -15,6 +15,8 @@ def array_from_env(key, default:)
 end
 
 GlassOctopus.run(app) do |config|
+  config.logger = Logger.new(STDOUT)
+
   config.adapter :ruby_kafka do |kafka|
     kafka.broker_list = array_from_env("KAFKA_BROKER_LIST", default: %w[localhost:9092])
     kafka.topic       = ENV.fetch("KAFKA_TOPIC", "mytopic")
